DHT read 관련 수정
* read 함수 인자가 &byte만 지원

+++ TWS.ino
@@ -21,8 +21,8 @@ typedef struct _StaticSensorValue {
   long time;
   int water;
   int sound;
-  int temp;
-  int humi;
+  byte temp;
+  byte humi;
 } StaticSensorValue;
 
 enum DynamicSensorKind {
